refactor: replaced TabWrapper with tab buttons component and adjust padding in secondary header
